requirements.txt
absl-py==0.8.0
astor==0.8.0
Click==7.0
Flask==1.1.1
gast==0.3.1
google-pasta==0.1.7
grpcio==1.23.0
gunicorn==19.9.0
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.17.2
Pillow==6.1.0
protobuf==3.9.1
PyYAML==5.1.2
scipy==1.3.1
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.15.6
wrapt==1.11.2
